pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compprog-pygame"
version = "0.1.0"
description = "Starter Pygame project with build tooling for executable export"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.5,<3",
    "pymunk>=7.0,<8",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "pytest>=8.0",
]

[project.scripts]
compprog-pygame = "compprog_pygame.game:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]